+++ NEWS.md
@@ -1,4 +1,5 @@
 # etnservice (development version)
+- Added `get_archival_data_uuid()` which fetches a table with UUID references to archival data files. These UUIDs can be used to construct paths to download the archival data CSV files hosted on lifewatch.com (#126).
 - The function `write_dwc()` has been updated in the `etn` package and is no longer needed in `etnservice`. (#141)
 - The `jsonlite` package is no longer a direct dependency, it is only used by a maintenance script. (#144)
 # etnservice 0.6.0